backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-APScheduler==1.13.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.36
marshmallow==3.22.0
python-dotenv==1.0.1
Werkzeug==3.0.6
Pillow==10.4.0
requests==2.32.3
qrcode==7.4.2
python-barcode==0.15.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
cryptography==43.0.3
gunicorn==23.0.0
celery==5.4.0
redis==5.2.0
